% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_govtrack.R
\name{.add_amp}
\alias{.add_amp}
\title{Intelligently separate params with ampersands}
\usage{
.add_amp(x)
}
\arguments{
\item{x}{Character string}
}
\value{
The input string concatenated with `&` if it does not end
  with `?`, otherwise the input string.
}
\description{
Intelligently separate params with ampersands
}
\keyword{internal}

