% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_govtrack.R
\name{.bad_filters}
\alias{.bad_filters}
\title{Identify improper filters}
\usage{
.bad_filters(res, filter)
}
\arguments{
\item{res}{Resource identifier}

\item{filter}{Named list of query filters}
}
\value{
Named vector the same length as \code{filter} identifying
improper filters with 1 and proper filters with 0.
}
\description{
Identify improper filters
}
\keyword{internal}

