% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_govtrack.R
\name{.create_query}
\alias{.create_query}
\title{Create API query string}
\usage{
.create_query(res_url, filter, sort, limit, offset, select)
}
\arguments{
\item{res_url}{Resource root URL}

\item{filter}{Named list of query filters}

\item{sort}{Vector of fields on which to sort}

\item{limit}{Maximum number of results to return}

\item{offset}{Pagination index of results}

\item{select}{Vector of fields to return}
}
\value{
Character string representing the constructed API query
}
\description{
Create API query string
}
\keyword{internal}

