% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_govtrack.R
\name{fetch_govtrack}
\alias{.bad_filters}
\alias{.create_query}
\alias{fetch_govtrack}
\title{Retrieve data from the GovTrack.us API}
\usage{
fetch_govtrack(res = "bill", filter, sort, limit, offset)

.bad_filters(res, filter)

.create_query(res_url, filter, sort, limit, offset)
}
\arguments{
\item{res}{Resource identifier}

\item{filter}{Named list of query filters}

\item{sort}{Vector of fields on which to sort}

\item{limit}{Maximum number of results to return}

\item{offset}{Pagination index of results}

\item{res_url}{Resource root URL}
}
\value{
Data frame of results. Note that observations are returned
  according to the structure returned by the server and may not be
  wholly atomic.

Named vector the same length as \code{filter} identifying
improper filters with 1 and proper filters with 0.

Character string representing the constructed API query
}
\description{
Queries a GovTrack.us API resource for a data set using server-side
filtering, sorting, and pagination.

Identify improper filters

Create API query string
}

