% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fields}
\alias{fields}
\title{Schemas of access points available at GovTrack.us}
\format{A data frame with 132 observations and 4 variables:
\describe{
  \item{resource}{Name of access point}
  \item{field}{Field name}
  \item{filterable}{Boolean value representing whether the field is
    filterable}
  \item{sortable}{Boolean value representing whether the field is
    sortable}
}}
\source{
\url{https://www.govtrack.us/developers/api}
}
\usage{
fields
}
\description{
A dataset containing partial data dictionaries of the eight API
resources accessible at GovTrack.us.
}
\keyword{datasets}

