% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resources}
\alias{resources}
\title{API access points available at GovTrack.us}
\format{A data frame with 8 observations and 3 variables:
\describe{
  \item{resource}{Name of access point}
  \item{description}{Description of the schema}
  \item{resource_url}{Root URL of access point}
}}
\source{
\url{https://www.govtrack.us/developers/api}
}
\usage{
resources
}
\description{
A dataset containing the name and other attributes of the eight API
resources accessible at GovTrack.us.
}
\keyword{datasets}

