% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_govtrack.R
\name{fetch_govtrack}
\alias{fetch_govtrack}
\title{Retrieve data from the GovTrack.us API}
\usage{
fetch_govtrack(res = "bill", filter, sort, limit, offset)
}
\arguments{
\item{res}{Resource identifier}

\item{filter}{Named list of query filters}

\item{sort}{Vector of fields on which to sort}

\item{limit}{Maximum number of results to return}

\item{offset}{Pagination index of results}
}
\value{
Data frame of results. Note that observations are returned
  according to the structure returned by the server and may not be
  wholly atomic.
}
\description{
Queries a GovTrack.us API resource for a data set using server-side
filtering, sorting, and pagination.
}

